refactor: 최근요청시간 변수명 수정, LatestHealthCheckCallLogDto 패키지 변경
